Move neighbour finding to base class
